requirements.txt
PyQt6==6.5.1
mysql-connector-python==9.2.0
selenium==4.10.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv